refactor: organise clients into folder
