+++ cmd/rhmas/kafka/create/create.go
@@ -9,7 +9,7 @@ import (
 	mas "github.com/bf2fc6cc711aee1a0c2a/cli/client/mas"
 	commonflags "github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhmas/flags"
 	"github.com/bf2fc6cc711aee1a0c2a/cli/cmd/rhmas/kafka/flags"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/sdk"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/api/rhmas"
 	"github.com/spf13/cobra"
 )
 
@@ -36,7 +36,7 @@ func runCreate(cmd *cobra.Command, _ []string) {
 	provider := commonflags.MustGetDefinedString(flags.FlagProvider, cmd.Flags())
 	multiAZ := commonflags.MustGetBool(flags.FlagMultiAZ, cmd.Flags())
 
-	client := sdk.BuildMasClient()
+	client := rhmas.BuildClient()
 
 	kafkaRequest := mas.KafkaRequest{Name: name, Region: region, CloudProvider: provider, MultiAz: multiAZ}
 	response, status, err := client.DefaultApi.ApiManagedServicesApiV1KafkasPost(context.Background(), true, kafkaRequest)
