+++ cmd/rhmas/kafka/status/status.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 
 	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/config"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/sdk"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/api/rhmas"
 	"github.com/spf13/cobra"
 )
 
@@ -34,7 +34,7 @@ func runStatus(cmd *cobra.Command, args []string) {
 		return
 	}
 
-	client := sdk.BuildMasClient()
+	client := rhmas.BuildClient()
 
 	res, status, err := client.DefaultApi.ApiManagedServicesApiV1KafkasIdGet(context.Background(), id)
 	if err != nil {
