+++ cmd/rhmas/kafka/use/use.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/config"
-	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/sdk"
+	"github.com/bf2fc6cc711aee1a0c2a/cli/pkg/api/rhmas"
 
 	"github.com/spf13/cobra"
 )
@@ -30,7 +30,7 @@ func runUse(cmd *cobra.Command, args []string) {
 
 	id := args[0]
 
-	client := sdk.BuildMasClient()
+	client := rhmas.BuildClient()
 
 	res, status, err := client.DefaultApi.ApiManagedServicesApiV1KafkasIdGet(context.Background(), id)
 	if err != nil {
